+++ TFLLib/ITflDataLibrarian.cs
@@ -248,14 +248,14 @@ namespace TFLLib
 
       void InsertPlayerGameMetric(string playerId, string gameCode,
          int projYDp, int YDp, int projYDr, int ydr,
-         int projTDp, int TDp, int projTDr, int tdr,
+         int projTDp, int TDp, decimal projTDr, int tdr,
          int projTDc, int TDc, int projYDc, int YDc,
          int projFG, int fg, int projPat, int pat
          );
 
       string UpdatePlayerGameMetric(string playerId, string gameCode,
          int projYDp, int YDp, int projYDr, int ydr,
-         int projTDp, int TDp, int projTDr, int tdr,
+         int projTDp, int TDp, decimal projTDr, int tdr,
          int projTDc, int TDc, int projYDc, int YDc,
          int projFG, int fg, int projPat, int pat
          );
